+++ lib/libchen/src/plat/win/sys_win.cpp
@@ -19,10 +19,23 @@ std::string sys::error()
 
 std::string sys::proc(int argc, const char *const argv[])
 {
-	CHAR buf[MAX_PATH] = { 0 };
+    CHAR buf[MAX_PATH] = { 0 };
 
-	if (::GetModuleFileName(NULL, buf, sizeof(buf)))
-		return std::string(buf);
-	else
-		return "";
+    if (::GetModuleFileName(NULL, buf, sizeof(buf)))
+        return std::string(buf);
+    else
+        return "";
+}
+
+int sys::pid()
+{
+    return ::GetCurrentProcessId();
+}
+
+bool sys::kill(int pid)
+{
+    HANDLE handle = ::OpenProcess(PROCESS_TERMINATE, FALSE, pid);
+    BOOL ret = ::TerminateProcess(handle, 0);
+    ::CloseHandle(handle);
+    return ret == TRUE;
 }